Added meta_request for Rails console integration
Go to the Chrome Web Store and install Rails Console. Just do it. Do it
now.

Now load a page when running in the development environment. Open the
Chrome dev console, go to “Rails”, and gaze in wonder at a breakdown of
response times, SQL queries performed, view rendering durations, logs
and errors.